Simplified Google Places token loading in index.html
Now the token is not loaded dinamically when the user is logged in, but rendered
by the server always. Token is displayed for everyone in the public map widget,
no need for secrecy anymore.

Users will need to refresh the page when the token changes. App will not
complain about delays in loading the google library anymore.